requirements.txt
PyQt6>=6.6.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
chromadb>=0.4.18
tiktoken>=0.5.1
sentence-transformers>=2.2.0
# API Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
